webdav: handle permission errors in fs.ReadDir
Before this commit, a permission error during Walk aborted walking and
returned 403 Forbidden. Now, permission errors are handled so that dirs
that cause them are simply skipped and not included in the returned
FileInfo slice.

The io/fs import was aliased to iofs so that it could be used in the
ReadDir method which has a receiver with the same name.